feat: Implement cookie handler
