+++ account-service/src/main/kotlin/com/falsy/account/AccountServiceApplication.kt
@@ -1,0 +1,11 @@
+package com.falsy.account
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+@SpringBootApplication
+class AuthServiceApplication
+
+fun main(args: Array<String>) {
+    runApplication<AuthServiceApplication>(*args)
+}
