+++ account-service/src/main/kotlin/com/falsy/account/model/dto/ChallengeDto.kt
@@ -1,0 +1,11 @@
+package com.falsy.account.model.dto
+
+import jakarta.validation.constraints.NotBlank
+
+/**
+ * auth-service
+ *
+ * @author uuhnaut69
+ *
+ */
+data class ChallengeDto(@NotBlank(message = "Address can not blank") val address: String)
