+++ account-service/src/main/kotlin/com/falsy/account/model/dto/NonceDto.kt
@@ -1,0 +1,9 @@
+package com.falsy.account.model.dto
+
+/**
+ * auth-service
+ *
+ * @author uuhnaut69
+ *
+ */
+data class NonceDto(val nonce: String)
