+++ account-service/src/main/kotlin/com/falsy/account/model/entity/Role.kt
@@ -1,0 +1,11 @@
+package com.falsy.account.model.entity
+
+/**
+ * auth-service
+ *
+ * @author uuhnaut69
+ *
+ */
+enum class Role {
+    ROLE_USER, ROLE_ADMIN
+}
