+++ account-service/src/main/resources/application.yaml
