+++ account-service/src/test/kotlin/com/falsy/account/AuthServiceApplicationTests.kt
@@ -1,0 +1,13 @@
+package com.falsy.account
+
+import org.junit.jupiter.api.Test
+import org.springframework.boot.test.context.SpringBootTest
+
+@SpringBootTest
+class AuthServiceApplicationTests {
+
+    @Test
+    fun contextLoads() {
+    }
+
+}
