--- auth-service/src/main/kotlin/com/falsy/authentication/AuthServiceApplication.kt
@@ -1,11 +1,0 @@
-package com.falsy.authentication
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.runApplication
-
-@SpringBootApplication
-class AuthServiceApplication
-
-fun main(args: Array<String>) {
-    runApplication<AuthServiceApplication>(*args)
-}
