--- auth-service/src/main/kotlin/com/falsy/authentication/api/advice/RestExceptionHandler.kt
@@ -1,15 +1,0 @@
-package com.falsy.authentication.api.advice
-
-import org.springframework.web.bind.annotation.RestControllerAdvice
-import org.springframework.web.reactive.result.method.annotation.ResponseEntityExceptionHandler
-
-/**
- * auth-service
- *
- * @author uuhnaut69
- *
- */
-@RestControllerAdvice
-class RestExceptionHandler : ResponseEntityExceptionHandler() {
-
-}
