--- auth-service/src/main/kotlin/com/falsy/authentication/config/RedisConfig.kt
@@ -1,24 +1,0 @@
-package com.falsy.authentication.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory
-import org.springframework.session.data.redis.config.annotation.web.server.EnableRedisWebSession
-
-
-/**
- * auth-service
- *
- * @author uuhnaut69
- *
- */
-@Configuration
-@EnableRedisWebSession
-class RedisConfig {
-
-    @Bean
-    fun redisConnectionFactory(): LettuceConnectionFactory {
-        return LettuceConnectionFactory()
-    }
-
-}
