--- auth-service/src/main/kotlin/com/falsy/authentication/config/security/siwe/SiweCredentials.kt
@@ -1,11 +1,0 @@
-package com.falsy.authentication.config.security.siwe
-
-import java.io.Serializable
-
-/**
- * auth-service
- *
- * @author uuhnaut69
- *
- */
-data class SiweCredentials(val message: String, val signature: String) : Serializable
