--- auth-service/src/main/kotlin/com/falsy/authentication/exception/UnauthorizedException.kt
@@ -1,14 +1,0 @@
-package com.falsy.authentication.exception
-
-import org.springframework.http.HttpStatus
-import org.springframework.web.bind.annotation.ResponseStatus
-
-/**
- * auth-service
- *
- * @author uuhnaut69
- *
- */
-@ResponseStatus(HttpStatus.UNAUTHORIZED)
-class UnauthorizedException(message: String?) : RuntimeException(message) {
-}
