--- auth-service/src/main/kotlin/com/falsy/authentication/model/dto/NonceDto.kt
@@ -1,9 +1,0 @@
-package com.falsy.authentication.model.dto
-
-/**
- * auth-service
- *
- * @author uuhnaut69
- *
- */
-data class NonceDto(val nonce: String)
