--- auth-service/src/main/kotlin/com/falsy/authentication/model/entity/Role.kt
@@ -1,11 +1,0 @@
-package com.falsy.authentication.model.entity
-
-/**
- * auth-service
- *
- * @author uuhnaut69
- *
- */
-enum class Role {
-    ROLE_USER, ROLE_ADMIN
-}
