--- auth-service/src/test/kotlin/com/falsy/authentication/AuthServiceApplicationTests.kt
@@ -1,13 +1,0 @@
-package com.falsy.authentication
-
-import org.junit.jupiter.api.Test
-import org.springframework.boot.test.context.SpringBootTest
-
-@SpringBootTest
-class AuthServiceApplicationTests {
-
-    @Test
-    fun contextLoads() {
-    }
-
-}
