+++ gateway/dev.env
@@ -1,5 +1,6 @@
 NODE_ENV=production
 PORT=4000
 AUTH_SERVICE=http://auth-service:8080
+REDIS_URL=redis://redis:6379
 DEBUG=1
 SESSION_SECRET="the secret must have length 32 or greater"
