+++ gateway/package.json
@@ -7,8 +7,8 @@
     "prestart": "pnpm run mesh:build && pnpm run build",
     "build": "tsc -p tsconfig.json",
     "build:watch": "tsc -p tsconfig.json -w",
-    "dev": "concurrently \"npx tsc --watch\" \"nodemon -q dist/src/index.js\"",
-    "start": "node dist/src/index.js"
+    "dev": "concurrently \"pnpm build:watch\" \"nodemon -q dist/src/main.js\"",
+    "start": "node dist/src/main.js"
   },
   "keywords": [
     "gateway"
@@ -29,12 +29,16 @@
     "@graphql-mesh/plugin-snapshot": "^0.93.1",
     "@graphql-mesh/runtime": "^0.46.23",
     "@graphql-mesh/transform-resolvers-composition": "^0.13.20",
+    "@mgcrea/fastify-session": "^1.1.0",
+    "@mgcrea/fastify-session-redis-store": "^1.0.0",
     "@whatwg-node/fetch": "^0.8.8",
     "dotenv": "^16.0.3",
     "fastify": "^4.17.0",
-    "graphql": "^16.6.0"
+    "graphql": "^16.6.0",
+    "ioredis": "^5.3.2"
   },
   "devDependencies": {
+    "@types/ioredis": "^5.0.0",
     "@types/node": "^18.16.2",
     "concurrently": "^8.0.1",
     "nodemon": "^2.0.22",
