+++ gateway/src/types/user.type.ts
@@ -1,0 +1,3 @@
+export interface User {
+  address: string;
+}
