+++ marketplace/schema.graphql
@@ -2,25 +2,9 @@ directive @globalOptions(endpoint: String, operationHeaders: ObjMap, queryParams
 
 directive @httpOperation(httpMethod: HTTPMethod, isBinary: Boolean, operationSpecificHeaders: ObjMap, path: String, queryParamArgMap: ObjMap, queryStringOptionsByParam: ObjMap, requestBaseBody: ObjMap) on FIELD_DEFINITION
 
-input AuthenticationDto_Input {
-  address: String!
-  message: String!
-  signature: String!
-}
-
-"""
-The `BigInt` scalar type represents non-fractional signed whole numeric values.
-"""
-scalar BigInt
-
 input ChallengeDto_Input {
   address: String!
 }
-
-"""
-A date-time string at UTC, such as 2007-12-03T10:15:30Z, compliant with the `date-time` format outlined in section 5.6 of the RFC 3339 profile of the ISO 8601 standard for representation of dates and times using the Gregorian calendar.
-"""
-scalar DateTime
 
 enum HTTPMethod {
   CONNECT
@@ -34,41 +18,13 @@ enum HTTPMethod {
   TRACE
 }
 
-"""
-The `JSON` scalar type represents JSON values as specified by [ECMA-404](http://www.ecma-international.org/publications/files/ECMA-ST/ECMA-404.pdf).
-"""
-scalar JSON
-
 type Mutation {
-  """
-  
-  >**Method**: `POST`
-  >**Base URL**: `{env.AUTH_SERVICE}`
-  >**Path**: `/auth/challenge`
-  Perform challenge nonce for authentication.
-  
-  """
+  """Perform challenge nonce for verifying account."""
   challenge(input: ChallengeDto_Input): ResponseString
+  login(input: VerifyAccountDto_Input!): String
 
-  """
-  
-  >**Method**: `POST`
-  >**Base URL**: `{env.AUTH_SERVICE}`
-  >**Path**: `/auth/login`
-  Perform login action.
-  
-  """
-  login(input: AuthenticationDto_Input): ResponseBoolean
-
-  """
-  
-  >**Method**: `POST`
-  >**Base URL**: `{env.AUTH_SERVICE}`
-  >**Path**: `/auth/logout`
-  Perform logout action.
-  
-  """
-  logout(webSession: WebSession_Input!): ResponseBoolean
+  """Perform verify account action."""
+  verifyAccount(input: VerifyAccountDto_Input): ResponseString
 }
 
 scalar ObjMap
@@ -77,36 +33,13 @@ type Query {
   dummy: String
 }
 
-type ResponseBoolean {
-  data: Boolean
-  status: Int!
-}
-
 type ResponseString {
   data: String
   status: Int!
 }
 
-input WebSession_Input {
-  attributes: JSON
-  creationTime: DateTime
-  expired: Boolean
-  id: String
-  lastAccessTime: DateTime
-  maxIdleTime: mutationInput_logout_webSession_maxIdleTime_Input
-  started: Boolean
-}
-
-input mutationInput_logout_webSession_maxIdleTime_Input {
-  nano: Int
-  negative: Boolean
-  seconds: BigInt
-  units: [mutationInput_logout_webSession_maxIdleTime_units_items_Input]
-  zero: Boolean
-}
-
-input mutationInput_logout_webSession_maxIdleTime_units_items_Input {
-  dateBased: Boolean
-  durationEstimated: Boolean
-  timeBased: Boolean
+input VerifyAccountDto_Input {
+  address: String!
+  message: String!
+  signature: String!
 }
